@@ -1,10 +1,13 @@
 import React, { Component } from "react";
-import Link from "next/link";
+import Link from 'next/link';
 import { connect } from "react-redux";
 import { formatMoney } from "../../utils";
 import { addCart } from "../../utils/cart";
+import DataContext from "../Context/DataContext";
 
 class Hero extends Component {
+  static contextType = DataContext; 
+
   constructor(props) {
     super();
     const { produto } = props;
@@ -97,6 +100,7 @@ class Hero extends Component {
       },
       true
     );
+    this.context.updateQtd(2)
   }
 
   renderDetalhes() {
@@ -109,8 +113,7 @@ class Hero extends Component {
         </div>
         <div className="categoria">
           <p>
-            Categoria:&nbsp;
-            <span className="categoria-link">{produto.categoryName}</span>
+            Categoria:&nbsp;<span className="categoria-link">{produto.categoryName}</span>
           </p>
         </div>
         <br />
