@@ -1,52 +1,51 @@
-import React, { Component } from 'react';
+import React, { Component } from "react";
 
-import Layout from '../components/Layout';
+import Layout from "../components/Layout";
 
-import Cabecalho from '../containers/Cabecalho';
-import Banners from '../containers/Banners';
-import Beneficios from '../containers/Beneficios';
-import ProdutosPaginaInicial from '../containers/Lista/ProdutosPaginaInicial';
-import Rodape from '../containers/Rodape';
+import Cabecalho from "../containers/Cabecalho";
+import Banners from "../containers/Banners";
+import Beneficios from "../containers/Beneficios";
+import ProdutosPaginaInicial from "../containers/Lista/ProdutosPaginaInicial";
+import Rodape from "../containers/Rodape";
 
-import initialize from '../utils/initialize';
-import callBaseData from '../utils/callBaseData';
-import { fetchData } from '../utils/fetch';
+import initialize from "../utils/initialize";
+import callBaseData from "../utils/callBaseData";
+import { fetchData } from "../utils/fetch";
 
-import { connect } from 'react-redux';
-import actions from '../redux/actions';
-
+import { connect } from "react-redux";
+import actions from "../redux/actions";
 
 class Index extends Component {
+  state = {
+    qtd: 0,
+  }
+
 
   static async getInitialProps(ctx) {
     // initialize(ctx);
-    return callBaseData([
-      actions.fetchProdutosPaginaInicial
-    ], ctx);
+    return callBaseData([actions.fetchProdutosPaginaInicial], ctx);
   }
-  
 
   async componentDidMount() {
     await this.props.getUser({ token: this.props.token });
   }
 
-
   render() {
     return (
-      <Layout title="Sid Surf Store | Estilo e bom gosto">
-        <Cabecalho  />
-        <Banners />
-        <Beneficios />
-        <ProdutosPaginaInicial />
-        <Rodape />
-      </Layout>
 
-    )
+        <Layout title="Sid Surf Store | Estilo e bom gosto">
+          <Cabecalho />
+          <Banners />
+          <Beneficios />
+          <ProdutosPaginaInicial />
+          <Rodape />
+        </Layout>
+    );
   }
 }
 
 const mapStateToProps = (state) => ({
-  token: state.auth.token
+  token: state.auth.token,
 });
 
 export default connect(mapStateToProps, actions)(Index);
